refactor: move AuthModal to ui directory and update import path in AuthGate
- Moved AuthModal component to the ui directory for better organization.
- Updated the import path in AuthGate to reflect the new location of AuthModal.
